+++ packages/api-client/src/conversation/data/ConversationAddPermissionUpdateData.ts
@@ -1,6 +1,6 @@
 /*
  * Wire
- * Copyright (C) 2018 Wire Swiss GmbH
+ * Copyright (C) 2025 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,6 +17,8 @@
  *
  */
 
-import {Availability} from '@wireapp/protocol-messaging';
+import {ADD_PERMISSION} from '../Conversation';
 
-export type AvailabilityType = Availability.Type;
+export interface ConversationAddPermissionUpdateData {
+  add_permission: ADD_PERMISSION;
+}