+++ packages/core/src/conversation/ConversationService/ConversationService.ts
@@ -17,7 +17,6 @@
  *
  */
 
-import {GenericMessage} from '@pydio/protocol-messaging';
 import {
   Conversation,
   DefaultConversationRoleName,
@@ -48,10 +47,9 @@ import {Decoder} from 'bazinga64';
 
 import {APIClient} from '@wireapp/api-client';
 import {LogFactory, TypedEventEmitter} from '@wireapp/commons';
+import {GenericMessage} from '@wireapp/protocol-messaging';
 
 import {
-  AddUsersFailure,
-  AddUsersFailureReasons,
   AddUsersParams,
   BaseCreateConversationResponse,
   KeyPackageClaimUser,
@@ -366,26 +364,7 @@ export class ConversationService extends TypedEventEmitter<Events> {
     groupId,
     conversationId,
   }: Required<AddUsersParams>): Promise<BaseCreateConversationResponse> {
-    const notMLSCapableUserIds: string[] = [];
-    // Failure object for users that are not MLS capable
-    const notMLSCapableUserFailure: AddUsersFailure = {
-      reason: AddUsersFailureReasons.NOT_MLS_CAPABLE,
-      users: [],
-    };
-
-    // There is no blocker for adding users to a conversation that are not MLS capable
-    // We just need to skip them when claiming key packages and add them to the failedToAdd list
-    for (const user of qualifiedUsers) {
-      const protocols = await this.apiClient.api.user.getUserSupportedProtocols(user);
-      if (!protocols.includes(ConversationProtocol.MLS)) {
-        notMLSCapableUserIds.push(user.id);
-        notMLSCapableUserFailure.users.push(user);
-      }
-    }
-
-    // We need to claim key packages only for users that are MLS capable
-    const mlsCapableUsers: QualifiedId[] = qualifiedUsers.filter(user => !notMLSCapableUserIds.includes(user.id));
-    const {keyPackages, failures: keysClaimingFailures} = await this.mlsService.getKeyPackagesPayload(mlsCapableUsers);
+    const {keyPackages, failures: keysClaimingFailures} = await this.mlsService.getKeyPackagesPayload(qualifiedUsers);
 
     await this.mlsService.addUsersToExistingConversation(groupId, keyPackages);
 
@@ -396,11 +375,7 @@ export class ConversationService extends TypedEventEmitter<Events> {
 
     return {
       conversation,
-      failedToAdd: [
-        ...keysClaimingFailures,
-        // In case we have users that are not MLS capable, we add them to the failedToAdd list
-        ...(notMLSCapableUserIds.length > 0 ? [notMLSCapableUserFailure] : []),
-      ],
+      failedToAdd: [...keysClaimingFailures, ...keysClaimingFailures],
     };
   }