+++ packages/react-ui-kit/src/Form/ButtonsGroup/ButtonsGroup.styles.ts
@@ -1,6 +1,6 @@
 /*
  * Wire
- * Copyright (C) 2018 Wire Swiss GmbH
+ * Copyright (C) 2025 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,6 +17,9 @@
  *
  */
 
-import {Availability} from '@wireapp/protocol-messaging';
+import {CSSObject} from '@emotion/react';
 
-export type AvailabilityType = Availability.Type;
+export const wrapperStyles: CSSObject = {
+  display: 'flex',
+  alignItems: 'center',
+};